requirements.txt
curl_cffi>=0.7.0
python-telegram-bot>=20.7
pyyaml>=6.0
httpx>=0.27.0
thefuzz>=0.22.0
python-Levenshtein>=0.25.0
